+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.9.13
 
 * Extend extension hat exmples to GPIO only.
-* Add Adafruit TSL2591 light sensor
+* Add Adafruit [TSL2591](https://github.com/pezi/dart_periphery/blob/main/example/i2c_tsl2591.dart) light sensor
 
 ## 0.9.12